Rename and refactor new importer
