+++ lib/tasks/reference_data.rake
@@ -25,4 +25,12 @@ namespace :reference_data do
   task create_assessment_questions: :environment do
     AssessmentQuestions::Importer.new.call
   end
+
+  desc 'create NOMIS alert mappings'
+  task create_nomis_alerts: :environment do
+    NomisAlerts::Importer.new(
+      alert_types: NomisClient::AlertTypes.get,
+      alert_codes: NomisClient::AlertCodes.get
+    ).call
+  end
 end
